Small update lesson9

+++ lesson_9/includes/modulation.h
@@ -1,6 +1,6 @@
 #define SAMPLES_ON_BIT 10
 
-double* BPSK_modulation(uint8_t* bits, int bits_count, int* symbols_count);
+int16_t* BPSK_modulation(uint8_t* bits, int bits_count, int* symbols_count);
 int* QPSK_modulation(uint8_t* bits, int bits_count, int* symbols_count);
 // int* upsampling(int* symbols, int symbols_count, int L, int* out_size);
-double* upsampling(double* symbols, int symbols_count, int L, int* out_size);
+int16_t* upsampling(int16_t* symbols, int symbols_count, int L, int* out_size);
